requirements.txt
matplotlib
pandas
scikit-learn
joblib
fastapi
uvicorn
python-dotenv